Add high_consequence flag to template
This commit introduces the high_consequence flag. This is intended
for vendors doing analysis to suggest that readers focus on a
particular vulnerability as it may be more exploitable or more
consequential than other vulnerabilities.

This does not increment the template version, as it adds a field
without modifying existing fields.